Remove unused open argument for Terminal::draw()
Not used, and it isn't clear the terminal itself would be able to close
the window the widget is hosted it?  Seems more reasonable to instead
defer that logic to whomever is doing the job control of the thing on
the other side of the console (e.g. parent of bash.exe).